axolotl-shiftos: charger: add custom animation
This is actually the default animation with percentage added.
Percentage is centered on the X axis but fixed to 1200 pixels on
the Y axis.

Soooo... we pretty much hardcoded it for our 2160x1080 SHIFT6mq.
We need to do better in future :P

At least we are using SHIFT Petrol (dark mode) accent color (#28BDD7)
as text color.

Change-Id: I7e634d5138920737618ae8b02785fe6accc39d92
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>